third_party/blink/web_tests/svg/dynamic-updates/SVGMaskElement-svgdom-x-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'x' property of the SVGMaskElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS maskElement.x.baseVal.valueAsString is "-100%"
PASS maskElement.x.baseVal.valueAsString is "-10%"
PASS successfullyParsed is true

TEST COMPLETE
